backend1/Pranav/app/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
python-dotenv 
imaplib2 
pdf2image 
paddleocr 
llama-index 
huggingface_hub 
jsonschema